move IAuthenticator into spi
