+++ broker/src/main/java/org/eclipse/moquette/server/Server.java
@@ -15,21 +15,22 @@
  */
 package org.eclipse.moquette.server;
 
-import org.eclipse.moquette.server.netty.NettyAcceptor;
-import org.eclipse.moquette.spi.impl.AcceptAllAuthenticator;
-import org.eclipse.moquette.spi.impl.FileAuthenticator;
-import org.eclipse.moquette.spi.impl.SimpleMessaging;
-import org.eclipse.moquette.spi.persistence.MapDBPersistentStore;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import static org.eclipse.moquette.commons.Constants.PASSWORD_FILE_PROPERTY_NAME;
+import static org.eclipse.moquette.commons.Constants.PERSISTENT_STORE_PROPERTY_NAME;
 
 import java.io.File;
 import java.io.IOException;
 import java.text.ParseException;
 import java.util.Properties;
 
-import static org.eclipse.moquette.commons.Constants.PASSWORD_FILE_PROPERTY_NAME;
-import static org.eclipse.moquette.commons.Constants.PERSISTENT_STORE_PROPERTY_NAME;
+import org.eclipse.moquette.server.netty.NettyAcceptor;
+import org.eclipse.moquette.spi.IAuthenticator;
+import org.eclipse.moquette.spi.impl.AcceptAllAuthenticator;
+import org.eclipse.moquette.spi.impl.FileAuthenticator;
+import org.eclipse.moquette.spi.impl.SimpleMessaging;
+import org.eclipse.moquette.spi.persistence.MapDBPersistentStore;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 /**
  * Launch a  configured version of the server.
  * @author andrea
