+++ broker/src/main/java/org/eclipse/moquette/spi/IAuthenticator.java
@@ -13,7 +13,7 @@
  *
  * You may elect to redistribute this code under either of these licenses.
  */
-package org.eclipse.moquette.server;
+package org.eclipse.moquette.spi;
 
 /**
  * username and password checker
