+++ broker/src/main/java/org/eclipse/moquette/spi/impl/AcceptAllAuthenticator.java
@@ -1,6 +1,6 @@
 package org.eclipse.moquette.spi.impl;
 
-import org.eclipse.moquette.server.IAuthenticator;
+import org.eclipse.moquette.spi.IAuthenticator;
 
 /**
  * Created by andrea on 8/23/14.
